requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
idna==3.4
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
PyJWT==2.6.0
pytest==7.2.1
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
stripe==5.2.0
urllib3==1.26.14